+++ significant-figures.cabal
@@ -13,7 +13,15 @@ extra-source-files: CHANGELOG.md
 library
   default-language: Haskell2010
   hs-source-dirs:   lib
-  exposed-modules:  SigFig
+  exposed-modules:  Data.SigFig
+                   ,Data.SigFig.Interface
+                   ,Data.SigFig.Evaluate
+                   ,Data.SigFig.Parse
+                   ,Data.SigFig.Types
+                   ,Data.SigFig.Util
+  other-extensions: OverloadedStrings
+                   ,BlockArguments
+                   ,ImportQualifiedPost
   build-depends:    base ^>=4.14.3.0
                    ,text
                    ,parsec
